Pagination

+++ controllers/admin.js
@@ -1,6 +1,6 @@
 const { validationResult } = require("express-validator");
 const Product = require("../models/product");
-const mongoose = require("mongoose");
+const fileHelper = require("../util/file");
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
@@ -18,7 +18,6 @@ exports.postAddProduct = (req, res, next) => {
   const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
-  console.log(999, image);
   if (!image) {
     return res.status(422).render("admin/edit-product", {
       pageTitle: "Add Product",
@@ -34,8 +33,7 @@ exports.postAddProduct = (req, res, next) => {
       validationErrors: [],
     });
   }
-  const imageUrl = image.path;
-  console.log(111, imageUrl);
+  const imageUrl = "/" + image.path;
 
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
@@ -147,7 +145,6 @@ exports.getProducts = (req, res, next) => {
     // .select('title price -_id')
     // .populate('userId', 'name')
     .then((products) => {
-      console.log(products);
       res.render("admin/products", {
         prods: products,
         pageTitle: "Admin Products",
@@ -157,12 +154,24 @@ exports.getProducts = (req, res, next) => {
     .catch((err) => console.log(err));
 };
 
-exports.postDeleteProduct = (req, res, next) => {
-  const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+exports.deleteProduct = (req, res, next) => {
+  const prodId = req.params.productId;
+  Product.findById(prodId)
+    .then((product) => {
+      console.log(888, product);
+      if (!product) {
+        return next(new Error("Product not found."));
+      }
+
+      fileHelper.deleteFile(product.imageUrl);
+
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log("DESTROYED PRODUCT");
-      res.redirect("/admin/products");
+      res.status(200).json({ message: "Success!" });
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      res.status(500).json({ message: "Deleting product failed." });
+    });
 };
